requirements.txt
# Core
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
scikit-image>=0.24
opencv-python>=4.9
joblib>=1.3
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8

# Experiment tracking
mlflow>=2.12
wandb>=0.17

# Monitoring
evidently>=0.4.29

# Serving
fastapi>=0.110
uvicorn[standard]>=0.27

# Dev
pytest>=8.0
ruff>=0.4

